Move cpp_function to its own header file for further refactors
